move initialize() to module_utils
